==> 01-scRNAseq Acute Wounds/03-Seurat/s3_Seurat_eachCluster_Top30genes.Rmd <==
---
title: "scRNAseq cell type annotation for each cluster"
author: "Zhuang Liu"
date: "13/10/3021 Updated: `r Sys.Date()`"
output:
  html_document:
    toc: true #add a table of contents (TOC)
    toc_depth: 3 #headers' level in the toc
    toc_float: true #toc visible even when scrolled
    theme: lumen
    highlight: tango
    code_folding: show
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, fig.align='center', fig.width=12, fig.height = 8)
```

```{r warning=FALSE, message=FALSE}
library(Seurat)
library(tidyverse)
library(magrittr)
library(ggplot2)
library(patchwork)
library(cowplot)
library(knitr)
library(kableExtra)
```

#####################
# Reload the data
#####################
```{r fig.width=6, fig.height=5}
hswound.combined.sct <- readRDS("s2_cleanSeurat_harmony_allSamples_clusters_normGenes.rds")
hswound.combined.sct
table(hswound.combined.sct$SCT_snn_res.0.8) #cell number for each cluster
table(hswound.combined.sct$orig.ident) #cell number for each sample
markers <- readxl::read_xlsx("s2_MarkerGenes_SingleR_Annotation_manually.xlsx", sheet = 1)
table(markers$cluster) #marker genes for each cluster
top30 <- markers %>% group_by(cluster) %>% top_n(n = 30, wt = avg_log2FC)
DimPlot(hswound.combined.sct, reduction = "umap", group.by='seurat_clusters', label=TRUE) +
  ggtitle('') + NoLegend() + NoAxes()
####----Create the unsupervised hierarchical tree----####
DefaultAssay(hswound.combined.sct) <- "SCT"
hswound.combined.sct <- BuildClusterTree(object = hswound.combined.sct, assay = "SCT", slot = "scale.data")
suppressMessages(library(scales))
cluster.color <- hue_pal()(length(levels(hswound.combined.sct)))
# pull the tree
hswound.clustertree <- Tool(object = hswound.combined.sct, slot = "BuildClusterTree")
# plot the tree
suppressMessages(require(ape))
{plot.phylo(x = hswound.clustertree, direction = "rightwards", use.edge.length = TRUE, font = 4, tip.color = cluster.color, edge.width = 2, label.offset = 5, no.margin = TRUE)
tiplabels(pch = 19, col = cluster.color, adj = 3, cex = 2)}

DefaultAssay(hswound.combined.sct) <- "RNA" #Go back to the RNA assay
```
```{r fig.height=10, fig.width=12}
#stacked barplot showing the proportion of cells from each batch in each cluster
####----Please pay attentions to the original numbers of cells per sample----####
clusters <- unique(hswound.combined.sct@meta.data[["seurat_clusters"]])
clusters <- clusters[order(clusters)]
df <- data.frame()
for(i in 1:length(clusters)){
  SmCell_sum <- table(hswound.combined.sct$orig.ident)
  tmp.df1 <- hswound.combined.sct@meta.data %>% subset(seurat_clusters == clusters[i]) %>% select(orig.ident) %>% table()
  if(length(tmp.df1) == 12){
    #First normalize the total cell counts per sample
    cur_df <- as.data.frame(tmp.df1 / SmCell_sum)
    colnames(cur_df) <- c("Sample", "Freq")
    #Calculate the normalized proportion 
    cur_df$Freq <- cur_df$Freq * 1/(sum(cur_df$Freq))
    cur_df$Cluster <- clusters[i]
    df <- rbind(df, cur_df)
  } else {
    #print(i);print(tmp.df1)
    #only include the matched samples
    match.sample <- SmCell_sum[names(SmCell_sum) %in% names(tmp.df1)]
    #print(match.sample)
    #First normalize the total cell counts per sample
    cur_df <- as.data.frame(tmp.df1 / match.sample)
    colnames(cur_df) <- c("Sample", "Freq")
    #Calculate the normalized proportion 
    cur_df$Freq <- cur_df$Freq * 1/(sum(cur_df$Freq))
    cur_df$Cluster <- clusters[i]
    df <- rbind(df, cur_df)
  }
}

##Plot for each sample
ggplot(df, aes(x = Cluster, y = Freq, fill = Sample)) +
  geom_bar(position = "stack", stat = "identity") +
  scale_fill_brewer(palette = "Paired") + # Paired or Set3 maximum variable = 12
  xlab('Clsuter') +
  scale_y_continuous(breaks = seq(0, 1, 0.1),
                     expand = c(0, 0),
                     name = 'Percentage') +
  theme_bw() +
  theme(
    panel.grid.major.x = element_blank(),
    panel.grid.major.y = element_blank(),
    panel.grid.minor = element_blank(),
    strip.background = element_blank(),
    #panel.border = element_rect(size = 0.35),
    axis.ticks = element_blank(),
    axis.text.x = element_text(size = 8)
  ) +
  coord_flip()

##Plot for each group
df.group <- df %>% mutate(Sample = gsub("PWH..", "", Sample)) %>% group_by(Cluster, Sample) %>% 
  summarise(Freq = sum(Freq))
df.group$Sample <- factor(df.group$Sample, levels = c("D0", "D1", "D7", "D30"))
ggplot(df.group, aes(x = Cluster, y = Freq, fill = Sample)) +
  geom_bar(position = "stack", stat = "identity") +
  scale_fill_brewer(palette = "Set1") + # Paired or Set3 maximum variable = 12
  xlab('Clsuter') +
  scale_y_continuous(breaks = seq(0, 1, 0.1),
                     expand = c(0, 0),
                     name = 'Percentage') +
  theme_bw() +
  theme(
    panel.grid.major.x = element_blank(),
    panel.grid.major.y = element_blank(),
    panel.grid.minor = element_blank(),
    strip.background = element_blank(),
    #panel.border = element_rect(size = 0.35),
    axis.ticks = element_blank(),
    axis.text.x = element_text(size = 8)
  ) +
  coord_flip()
```


# 1. Keratinocytes
## 1.1 Basal keratinocytes: KRT5, KRT14
```{r fig.align='center', fig.width=8, fig.height=4}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c("KRT5", "KRT14"),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 2)
```

### 1.1.1 Cluster 2
```{r fig.align='center', fig.width=12, fig.height=8}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c(top30[top30$cluster == 2, ]$gene[1:12]),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
top30[top30$cluster == 2, ] %>%
  kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```

### 1.1.2 Cluster 20
```{r fig.align='center', fig.width=12, fig.height=8}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c(top30[top30$cluster == 20, ]$gene[1:12]),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
top30[top30$cluster == 20, ] %>%
  kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```

### 1.1.3 Cluster 3
```{r fig.align='center', fig.width=12, fig.height=8}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c(top30[top30$cluster == 3, ]$gene[1:12]),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
top30[top30$cluster == 3, ] %>%
  kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```

### 1.1.4 Cluster 17
```{r fig.align='center', fig.width=12, fig.height=8}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c(top30[top30$cluster == 17, ]$gene[1:12]),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
top30[top30$cluster == 17, ] %>%
  kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```

### 1.1.5 Cluster 11
```{r fig.align='center', fig.width=12, fig.height=8}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c(top30[top30$cluster == 11, ]$gene[1:12]),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
top30[top30$cluster == 11, ] %>%
  kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```

## 1.2 Suprabasal keratinocytes: KRT5, KRT14, KRT1, KRT10
```{r fig.align='center', fig.width=8, fig.height=2}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c("KRT5", "KRT14", "KRT1", "KRT10"),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
```

### 1.2.1 Cluster 4
```{r fig.align='center', fig.width=12, fig.height=8}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c(top30[top30$cluster == 4, ]$gene[1:12]),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
top30[top30$cluster == 4, ] %>%
  kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```

### 1.2.2 Cluster 15
```{r fig.align='center', fig.width=12, fig.height=8}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c(top30[top30$cluster == 15, ]$gene[1:12]),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
top30[top30$cluster == 15, ] %>%
  kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```

## 1.3 Suprabasal keratinocytes: KRT1, KRT10
```{r fig.align='center', fig.width=8, fig.height=4}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c("KRT1", "KRT10"),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 2)
```

### 1.3.1 Cluster 0
```{r fig.align='center', fig.width=12, fig.height=8}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c(top30[top30$cluster == 0, ]$gene[1:12]),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
top30[top30$cluster == 0, ] %>%
  kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```

### 1.3.2 Cluster 14
```{r fig.align='center', fig.width=12, fig.height=8}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c(top30[top30$cluster == 14, ]$gene[1:12]),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
top30[top30$cluster == 14, ] %>%
  kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```


# 2. Fibroblasts
## Broad fibroblasts markers: COL1A1, COL1A2, LUM
```{r fig.align='center', fig.width=8, fig.height=2}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c('COL1A1', 'COL1A2', 'LUM'),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
```

## 2.1 Cluster 7
```{r fig.align='center', fig.width=12, fig.height=8}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c(top30[top30$cluster == 7, ]$gene[1:12]),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
top30[top30$cluster == 7, ] %>%
  kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```

## 2.2 Cluster 8
```{r fig.align='center', fig.width=12, fig.height=8}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c(top30[top30$cluster == 8, ]$gene[1:12]),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
top30[top30$cluster == 8, ] %>%
  kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```

## 2.3 Cluster 1
```{r fig.align='center', fig.width=12, fig.height=8}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c(top30[top30$cluster == 1, ]$gene[1:12]),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
top30[top30$cluster == 1, ] %>%
  kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```

## 2.4 Cluster 12
```{r fig.align='center', fig.width=12, fig.height=8}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c(top30[top30$cluster == 12, ]$gene[1:12]),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
top30[top30$cluster == 12, ] %>%
  kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```

## 2.6 Known marker for subfibroblasts
```{r fig.align='center', fig.width=12, fig.height=8}
##mesenchymal fibroblasts: ASPN, POSTN, COMP
##secretory papillary fibroblasts: APCDD1, COL18A1, COL13A1
##pro-inflammatory fibroblasts: APOE, CCL19, CXCL3
##secretory reticular fibroblasts: WISP2, ANGPTL1, MFAP5
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c("ASPN", "COL23A1", "APOE", "ANGPTL1", "POSTN", "COL18A1", "CCL19", "MFAP5", "COMP", "COL13A1", "CXCL3"),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
```


# 3. Pericytes & SMCs
## Cluster 18
```{r fig.align='center', fig.width=12, fig.height=8}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c(top30[top30$cluster == 18, ]$gene[1:12]),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
top30[top30$cluster == 18, ] %>%
  kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```


# 4. Melanocytes
## Cluster 19
```{r fig.align='center', fig.width=12, fig.height=8}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c(top30[top30$cluster == 19, ]$gene[1:12]),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
top30[top30$cluster == 19, ] %>%
  kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```


# 5. Mast cells
## Cluster 10
```{r fig.align='center', fig.width=12, fig.height=8}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c(top30[top30$cluster == 10, ]$gene[1:12]),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
top30[top30$cluster == 10, ] %>%
  kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```


# 6. T cells
## Broad T-cell markers: CD3D, CD3E, CD2, CD7
```{r fig.align='center', fig.width=8, fig.height=2}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c('CD3D', 'CD3E', 'CD2', 'CD7'),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
```

## 6.1 Cluster 6
```{r fig.align='center', fig.width=12, fig.height=8}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c(top30[top30$cluster == 6, ]$gene[1:12]),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
top30[top30$cluster == 6, ] %>%
  kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```

## 6.2 Cluster 16
```{r fig.align='center', fig.width=12, fig.height=8}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c(top30[top30$cluster == 16, ]$gene[1:12]),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
top30[top30$cluster == 16, ] %>%
  kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```


# 7. Neurons
## Cluster 26
```{r fig.align='center', fig.width=12, fig.height=8}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c(top30[top30$cluster == 26, ]$gene[1:12]),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
top30[top30$cluster == 26, ] %>%
  kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```


# 8. B/plasma cells
## Broad B-cell markers: MS4A1, CD79A, CD74
```{r fig.align='center', fig.width=8, fig.height=2}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c('MS4A1', 'CD79A'),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
```

## Cluster 22
```{r fig.align='center', fig.width=12, fig.height=8}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c(top30[top30$cluster == 22, ]$gene[1:12]),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
top30[top30$cluster == 22, ] %>%
  kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```


# 9. Myeloid cells
## Broad myeloid cells markers: LYZ, MHC class II genes
```{r fig.align='center', fig.width=8, fig.height=2}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c('LYZ', 'HLA-DRB1', 'HLA-DRA', 'HLA-DQB2'),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
```

## 9.1 Cluster 21
```{r fig.align='center', fig.width=12, fig.height=8}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c(top30[top30$cluster == 21, ]$gene[1:12]),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
top30[top30$cluster == 21, ] %>%
  kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```

## 9.2 Cluster 9
```{r fig.align='center', fig.width=12, fig.height=8}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c(top30[top30$cluster == 9, ]$gene[1:12]),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
top30[top30$cluster == 9, ] %>%
  kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```

## 9.3 Cluster 24
```{r fig.align='center', fig.width=12, fig.height=8}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c(top30[top30$cluster == 24, ]$gene[1:12]),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
top30[top30$cluster == 24, ] %>%
  kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```

## 9.4 Cluster 5
```{r fig.align='center', fig.width=12, fig.height=8}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c(top30[top30$cluster == 5, ]$gene[1:12]),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
top30[top30$cluster == 5, ] %>%
  kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```

## 9.5 Cluster 25
```{r fig.align='center', fig.width=12, fig.height=8}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c(top30[top30$cluster == 25, ]$gene[1:12]),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
top30[top30$cluster == 25, ] %>%
  kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```


# 10. Endothelial cells
## Broad endothelial cells markers: CLDN5, VWF, PECAM1
```{r fig.align='center', fig.width=8, fig.height=2}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c('CLDN5', 'VWF', 'PECAM1'),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
```

## 10.1 Cluster 13
```{r fig.align='center', fig.width=12, fig.height=8}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c(top30[top30$cluster == 13, ]$gene[1:12]),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
top30[top30$cluster == 13, ] %>%
  kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```

## 10.2 Cluster 23
```{r fig.align='center', fig.width=12, fig.height=8}
p2 <- FeaturePlot(
  hswound.combined.sct,
  features = c(top30[top30$cluster == 23, ]$gene[1:12]),
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p2, function(x) {x + NoLegend() + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 4)
top30[top30$cluster == 23, ] %>%
  kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```


```{r eval=FALSE, fig.width=3, fig.height=2.5}
FeaturePlot(
  hswound.combined.sct,
  features = c('CD86'), #Check any marker gene
  cols = c("gray","red")) + NoLegend() + NoAxes() + theme(text = element_text(size = 10))
```

```{r fig.width=12}
anno <- data.table::fread("cell_identity_MariaKasper_Lab.txt", header = FALSE)
colnames(anno) <- c("barcode", "cellType")
metadata <- hswound.combined.sct@meta.data %>% rownames_to_column("barcode") %>% left_join(., anno, by=c("barcode" = "barcode"))

hswound.combined.sct <- AddMetaData(hswound.combined.sct, metadata$cellType, col.name = "celltype")

(DimPlot(hswound.combined.sct, reduction = "umap", group.by='seurat_clusters', label=TRUE) +
  ggtitle('') + NoLegend() + NoAxes()) +
  (DimPlot(hswound.combined.sct, reduction = "umap", group.by='celltype', label=TRUE) +
  ggtitle('') + NoLegend() + NoAxes())
```


# SessionInfo
```{r}
sessionInfo()
```
