==> 01-scRNAseq Acute Wounds/03-Seurat/s5_GOanalysis_eachCluster.Rmd <==
---
title: "GO results for each cluster"
author: "Zhuang Liu"
date: "20/10/3021 Updated: `r Sys.Date()`"
output:
  html_document:
    toc: true #add a table of contents (TOC)
    toc_depth: 3 #headers' level in the toc
    toc_float: true #toc visible even when scrolled
    theme: lumen
    highlight: tango
    code_folding: show
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, fig.align='center', fig.width=12, fig.height = 8, warning=FALSE, message=FALSE)
```

```{r warning=FALSE, message=FALSE}
library(tidyverse)
library(magrittr)
library(ggplot2)
library(patchwork)
library(cowplot)
library(knitr)
library(kableExtra)
```

###############################
# GO analysis for each cluster
###############################
```{r warning=FALSE, message=FALSE}
GO_BP <- function(i=NULL){
  datatmp <- markers %>% filter(CellType == clustername[i])
  require(clusterProfiler) #loading the package if it is not loaded
  require(org.Hs.eg.db)
  Enriched_gene_up <- as.character(unique(datatmp$entrezgene_id[1:200])[1:100])
  ego_BP_up <- enrichGO(gene = Enriched_gene_up,
                        OrgDb         = org.Hs.eg.db,
                        ont           = "BP",
                        keyType = "ENTREZID",
                        pAdjustMethod = "BH",
                        pvalueCutoff  = 1,
                        qvalueCutoff  = 1,
                        minGSSize = 10,
                        maxGSSize = 500,
                        readable = T)
  dotplot(ego_BP_up, showCategory=12, label_format = 100) + ggtitle(paste0("Cluster ", match(clustername[i], clusters)-1, ":", clustername[i], "-GO_BP")) + 
    theme(
      #panel.grid.major.x = element_blank(),
      #panel.grid.major.y = element_blank(),
      panel.grid.minor = element_blank(),
      panel.border = element_rect(size = 0.5, color = "black"),
      axis.text.x = element_text(size = 16, color = "black"),
      axis.text.y = element_text(size = 16, color = "black"))
}
KEGG <- function(i=NULL){
  datatmp <- markers %>% filter(CellType == clustername[i])
  require(clusterProfiler) #loading the package if it is not loaded
  require(org.Hs.eg.db)
  Enriched_gene_up <- as.character(unique(datatmp$entrezgene_id[1:200])[1:100])
  kk_up <- enrichKEGG(gene         = Enriched_gene_up,
                      organism     = 'hsa',
                      pAdjustMethod = "BH",
                      pvalueCutoff = 1,
                      qvalueCutoff  = 1,
                      minGSSize = 10,
                      maxGSSize = 500)
  dotplot(kk_up, showCategory=12, label_format = 100) + ggtitle(paste0("Cluster ", match(clustername[i], clusters)-1, ":", clustername[i], "-KEGG")) + 
    theme(
      #panel.grid.major.x = element_blank(),
      #panel.grid.major.y = element_blank(),
      panel.grid.minor = element_blank(),
      panel.border = element_rect(size = 0.5, color = "black"),
      axis.text.x = element_text(size = 16, color = "black"),
      axis.text.y = element_text(size = 16, color = "black"))
}
```
```{r warning=FALSE, message=FALSE}
####----Load the marker genes----####
markers <- readxl::read_xlsx("s2_MarkerGenes_SingleR_Annotation_manually.xlsx", sheet = 1)
anno <- readxl::read_xlsx("s5_cell_types_for_each_cluster_manually_updated.xlsx", sheet = 1)

markers <- markers %>% left_join(., anno[,1:2], by=c("cluster" = "Cluster"))
table(markers$cluster)
clusters <- unique(markers$CellType)
clustername <- c("Basal-I", "Basal-II", "Basal-III", "Basal-IV",
                      "Spinous-I", "Spinous-II", "Spinous-III",
                      "Granular-I", "Granular-II",
                      "FB-I", "FB-II", "FB-III", "FB-IV",
                      "PC-vSMC", "MEL", "Schwann", 
                      "LE", "VE", 
                      "Mast-cell", 
                      "NK-cell", "Th", 
                      "Plasma_B-cell", 
                      "Mono-Mac",
                      "LC", 
                      "DC", "Dermal-DC", "DC-LAMP")
```

## `r paste0("Cluster ", match(clustername[1], clusters) -1, ":", clustername[1])`
```{r}
####----Print all the marker genes for each cluster----####
markers[markers$CellType == clustername[1], ] %>%
  slice_head(n = 100) %>% kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```
```{r fig.height=5, fig.width=12}
GO_BP(i=1)
```
```{r fig.height=5, fig.width=12}
KEGG(i=1)
```

## `r paste0("Cluster ", match(clustername[2], clusters) -1, ":", clustername[2])`
```{r}
####----Print all the marker genes for each cluster----####
markers[markers$CellType == clustername[2], ] %>%
  slice_head(n = 100) %>% kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```
```{r fig.height=5, fig.width=12}
GO_BP(i=2)
```
```{r fig.height=5, fig.width=12}
KEGG(i=2)
```

## `r paste0("Cluster ", match(clustername[3], clusters) -1, ":", clustername[3])`
```{r}
####----Print all the marker genes for each cluster----####
markers[markers$CellType == clustername[3], ] %>%
  slice_head(n = 100) %>% kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```
```{r fig.height=5, fig.width=12}
GO_BP(i=3)
```
```{r fig.height=5, fig.width=12}
KEGG(i=3)
```

## `r paste0("Cluster ", match(clustername[4], clusters) -1, ":", clustername[4])`
```{r}
####----Print all the marker genes for each cluster----####
markers[markers$CellType == clustername[4], ] %>%
  slice_head(n = 100) %>% kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```
```{r fig.height=5, fig.width=12}
GO_BP(i=4)
```
```{r fig.height=5, fig.width=12}
KEGG(i=4)
```

## `r paste0("Cluster ", match(clustername[5], clusters) -1, ":", clustername[5])`
```{r}
####----Print all the marker genes for each cluster----####
markers[markers$CellType == clustername[5], ] %>%
  slice_head(n = 100) %>% kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```
```{r fig.height=5, fig.width=12}
GO_BP(i=5)
```
```{r fig.height=5, fig.width=12}
KEGG(i=5)
```

## `r paste0("Cluster ", match(clustername[6], clusters) -1, ":", clustername[6])`
```{r}
####----Print all the marker genes for each cluster----####
markers[markers$CellType == clustername[6], ] %>%
  slice_head(n = 100) %>% kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```
```{r fig.height=5, fig.width=12}
GO_BP(i=6)
```
```{r fig.height=5, fig.width=12}
KEGG(i=6)
```

## `r paste0("Cluster ", match(clustername[7], clusters) -1, ":", clustername[7])`
```{r}
####----Print all the marker genes for each cluster----####
markers[markers$CellType == clustername[7], ] %>%
  slice_head(n = 100) %>% kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```
```{r fig.height=5, fig.width=12}
GO_BP(i=7)
```
```{r fig.height=5, fig.width=12}
KEGG(i=7)
```

## `r paste0("Cluster ", match(clustername[8], clusters) -1, ":", clustername[8])`
```{r}
####----Print all the marker genes for each cluster----####
markers[markers$CellType == clustername[8], ] %>%
  slice_head(n = 100) %>% kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```
```{r fig.height=5, fig.width=12}
GO_BP(i=8)
```
```{r fig.height=5, fig.width=12}
KEGG(i=8)
```

## `r paste0("Cluster ", match(clustername[9], clusters) -1, ":", clustername[9])`
```{r}
####----Print all the marker genes for each cluster----####
markers[markers$CellType == clustername[9], ] %>%
  slice_head(n = 100) %>% kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```
```{r fig.height=5, fig.width=12}
GO_BP(i=9)
```
```{r fig.height=5, fig.width=12}
KEGG(i=9)
```

## `r paste0("Cluster ", match(clustername[10], clusters) -1, ":", clustername[10])`
```{r}
####----Print all the marker genes for each cluster----####
markers[markers$CellType == clustername[10], ] %>%
  slice_head(n = 100) %>% kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```
```{r fig.height=5, fig.width=12}
GO_BP(i=10)
```
```{r fig.height=5, fig.width=12}
KEGG(i=10)
```

## `r paste0("Cluster ", match(clustername[11], clusters) -1, ":", clustername[11])`
```{r}
####----Print all the marker genes for each cluster----####
markers[markers$CellType == clustername[11], ] %>%
  slice_head(n = 100) %>% kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```
```{r fig.height=5, fig.width=12}
GO_BP(i=11)
```
```{r fig.height=5, fig.width=12}
KEGG(i=11)
```

## `r paste0("Cluster ", match(clustername[12], clusters) -1, ":", clustername[12])`
```{r}
####----Print all the marker genes for each cluster----####
markers[markers$CellType == clustername[12], ] %>%
  slice_head(n = 100) %>% kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```
```{r fig.height=5, fig.width=12}
GO_BP(i=12)
```
```{r fig.height=5, fig.width=12}
KEGG(i=12)
```

## `r paste0("Cluster ", match(clustername[13], clusters) -1, ":", clustername[13])`
```{r}
####----Print all the marker genes for each cluster----####
markers[markers$CellType == clustername[13], ] %>%
  slice_head(n = 100) %>% kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```
```{r fig.height=5, fig.width=12}
GO_BP(i=13)
```
```{r fig.height=5, fig.width=12}
KEGG(i=13)
```

## `r paste0("Cluster ", match(clustername[14], clusters) -1, ":", clustername[14])`
```{r}
####----Print all the marker genes for each cluster----####
markers[markers$CellType == clustername[14], ] %>%
  slice_head(n = 100) %>% kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```
```{r fig.height=5, fig.width=12}
GO_BP(i=14)
```
```{r fig.height=5, fig.width=12}
KEGG(i=14)
```

## `r paste0("Cluster ", match(clustername[15], clusters) -1, ":", clustername[15])`
```{r}
####----Print all the marker genes for each cluster----####
markers[markers$CellType == clustername[15], ] %>%
  slice_head(n = 100) %>% kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```
```{r fig.height=5, fig.width=12}
GO_BP(i=15)
```
```{r fig.height=5, fig.width=12}
KEGG(i=15)
```

## `r paste0("Cluster ", match(clustername[16], clusters) -1, ":", clustername[16])`
```{r}
####----Print all the marker genes for each cluster----####
markers[markers$CellType == clustername[16], ] %>%
  slice_head(n = 100) %>% kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```
```{r fig.height=5, fig.width=12}
GO_BP(i=16)
```
```{r fig.height=5, fig.width=12}
KEGG(i=16)
```

## `r paste0("Cluster ", match(clustername[17], clusters) -1, ":", clustername[17])`
```{r}
####----Print all the marker genes for each cluster----####
markers[markers$CellType == clustername[17], ] %>%
  slice_head(n = 100) %>% kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```
```{r fig.height=5, fig.width=12}
GO_BP(i=17)
```
```{r fig.height=5, fig.width=12}
KEGG(i=17)
```

## `r paste0("Cluster ", match(clustername[18], clusters) -1, ":", clustername[18])`
```{r}
####----Print all the marker genes for each cluster----####
markers[markers$CellType == clustername[18], ] %>%
  slice_head(n = 100) %>% kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```
```{r fig.height=5, fig.width=12}
GO_BP(i=18)
```
```{r fig.height=5, fig.width=12}
KEGG(i=18)
```

## `r paste0("Cluster ", match(clustername[19], clusters) -1, ":", clustername[19])`
```{r}
####----Print all the marker genes for each cluster----####
markers[markers$CellType == clustername[19], ] %>%
  slice_head(n = 100) %>% kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```
```{r fig.height=5, fig.width=12}
GO_BP(i=19)
```
```{r fig.height=5, fig.width=12}
KEGG(i=19)
```

## `r paste0("Cluster ", match(clustername[20], clusters) -1, ":", clustername[20])`
```{r}
####----Print all the marker genes for each cluster----####
markers[markers$CellType == clustername[20], ] %>%
  slice_head(n = 100) %>% kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```
```{r fig.height=5, fig.width=12}
GO_BP(i=20)
```
```{r fig.height=5, fig.width=12}
KEGG(i=20)
```

## `r paste0("Cluster ", match(clustername[21], clusters) -1, ":", clustername[21])`
```{r}
####----Print all the marker genes for each cluster----####
markers[markers$CellType == clustername[21], ] %>%
  slice_head(n = 100) %>% kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```
```{r fig.height=5, fig.width=12}
GO_BP(i=21)
```
```{r fig.height=5, fig.width=12}
KEGG(i=21)
```

## `r paste0("Cluster ", match(clustername[22], clusters) -1, ":", clustername[22])`
```{r}
####----Print all the marker genes for each cluster----####
markers[markers$CellType == clustername[22], ] %>%
  slice_head(n = 100) %>% kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```
```{r fig.height=5, fig.width=12}
GO_BP(i=22)
```
```{r fig.height=5, fig.width=12}
KEGG(i=22)
```

## `r paste0("Cluster ", match(clustername[23], clusters) -1, ":", clustername[23])`
```{r}
####----Print all the marker genes for each cluster----####
markers[markers$CellType == clustername[23], ] %>%
  slice_head(n = 100) %>% kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```
```{r fig.height=5, fig.width=12}
GO_BP(i=23)
```
```{r fig.height=5, fig.width=12}
KEGG(i=23)
```

## `r paste0("Cluster ", match(clustername[24], clusters) -1, ":", clustername[24])`
```{r}
####----Print all the marker genes for each cluster----####
markers[markers$CellType == clustername[24], ] %>%
  slice_head(n = 100) %>% kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```
```{r fig.height=5, fig.width=12}
GO_BP(i=24)
```
```{r fig.height=5, fig.width=12}
KEGG(i=24)
```

## `r paste0("Cluster ", match(clustername[25], clusters) -1, ":", clustername[25])`
```{r}
####----Print all the marker genes for each cluster----####
markers[markers$CellType == clustername[25], ] %>%
  slice_head(n = 100) %>% kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```
```{r fig.height=5, fig.width=12}
GO_BP(i=25)
```
```{r fig.height=5, fig.width=12}
KEGG(i=25)
```

## `r paste0("Cluster ", match(clustername[26], clusters) -1, ":", clustername[26])`
```{r}
####----Print all the marker genes for each cluster----####
markers[markers$CellType == clustername[26], ] %>%
  slice_head(n = 100) %>% kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```
```{r fig.height=5, fig.width=12}
GO_BP(i=26)
```
```{r fig.height=5, fig.width=12}
KEGG(i=26)
```

## `r paste0("Cluster ", match(clustername[27], clusters) -1, ":", clustername[27])`
```{r}
####----Print all the marker genes for each cluster----####
markers[markers$CellType == clustername[27], ] %>%
  slice_head(n = 100) %>% kbl() %>% 
  kable_material_dark() %>% 
  row_spec(0, bold = T, color = "black") %>%
  scroll_box(width = "100%", height = "400px")
```
```{r fig.height=5, fig.width=12}
GO_BP(i=27)
```
```{r fig.height=5, fig.width=12}
KEGG(i=27)
```

# SessionInfo
```{r}
sessionInfo()
```
